requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.6.3
apturl==0.5.2
argcomplete==1.12.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.0
bcrypt==3.1.7
bech32==1.2.0
bip32utils==0.3.post4
bitarray==1.2.2
blinker==1.4
boltons==21.0.0
Brlapi==0.7.0
Brownie==0.5.1
cachetools==4.2.2
certifi==2019.11.28
chardet==3.0.4
cliapp==1.20180812.1
click==8.0.3
cmdtest==0.32+git
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
cytoolz==0.11.0
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.0
defer==1.0.6
defusedxml==0.7.1
discord==1.7.3
discord-webhook==0.15.0
discord.py==1.7.3
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
ecdsa==0.16.1
entrypoints==0.3
et-xmlfile==1.1.0
eth-abi==2.1.1
eth-account==0.5.5
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
fasteners==0.14.1
Flask==2.0.2
fonttools==4.28.5
furl==2.1.3
future==0.18.2
h11==0.13.0
hexbytes==0.2.2
httplib2==0.14.0
idna==2.8
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
ipykernel==6.6.0
ipython==7.30.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonpickle==2.0.0
jsonschema==3.2.0
jupyter-client==7.1.0
jupyter-core==4.9.1
keyring==18.0.1
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
loguru==0.5.3
louis==3.12.0
lru-dict==1.1.7
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mnemonic==0.19
monotonic==1.5
multiaddr==0.0.9
multidict==5.1.0
nest-asyncio==1.5.4
netaddr==0.8.0
netifaces==0.10.4
networkx==2.6.3
numpy==1.21.4
oauthlib==3.1.0
odfpy==1.4.1
olefile==0.46
openpyxl==3.0.9
orderedmultidict==1.0.1
outcome==1.1.0
packaging==21.3
pandas==1.3.4
paramiko==2.6.0
parsimonious==0.8.1
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pipx==0.16.4
pluggy==1.0.0
prompt-toolkit==3.0.24
protobuf==3.18.0
py==1.11.0
py-solc-x==1.1.1
pycairo==1.16.2
pycodestyle==2.8.0
pycryptodome==3.10.1
pycups==1.9.73
Pygments==2.10.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.6
pyRFC3339==1.1
pyrsistent==0.18.0
pytest==6.2.5
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.19.2
python-telegram-bot==13.8.1
pytz==2019.3
pytz-deprecation-shim==0.1.0.post0
pyvis==0.1.9
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rlp==2.0.1
scikit-learn==1.0.1
scipy==1.7.3
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
selenium==4.1.0
semantic-version==2.8.5
simplejson==3.16.0
six==1.14.0
sklearn==0.0
sniffio==1.2.0
solc-select==0.2.1
sortedcontainers==2.4.0
systemd-python==234
terra-sdk==1.0.2
threadpoolctl==3.0.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
truffle==0.1.dev0
ttystatus==0.38
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
ubuntu-advantage-tools==27.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.8
usb-creator==0.3.7
userpath==1.7.0
varint==1.0.2
wadllib==1.3.3
wcwidth==0.2.5
web3==5.23.1
websockets==9.1
Werkzeug==2.0.2
Whoosh==2.7.4
wrapt==1.13.3
wsproto==1.0.0
xkit==0.0.0
yarl==1.6.3
